requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-mail-templated==2.6.5
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
reportlab==4.0.4
sqlparse==0.4.4
typing_extensions==4.7.1
XlsxWriter==3.1.2
